Address review: move device fixtures to top-level conftest
- Move any_device/gpu_device fixtures to src/tests/conftest.py so
  they're available to all tests (audio, text, video)
- Remove duplicate device detection from audio/conftest.py and
  text/embeddings_extraction_test.py
- Fix type hints: List[DeviceType] instead of bare list
- Fix SPARC missing-dep test to use any_device (doesn't need GPU)
