+++ src/tests/audio/conftest.py
@@ -112,47 +112,3 @@ def is_device_available() -> Callable[[DeviceType], bool]:
         return True  # CPU is always available
 
     return _is_device_available
-
-
-# ---------------------------------------------------------------------------
-# Device parameterization fixtures
-# ---------------------------------------------------------------------------
-def _available_devices() -> list:
-    """Detect available compute devices."""
-    devices = [DeviceType.CPU]
-    if torch.cuda.is_available():
-        devices.append(DeviceType.CUDA)
-    if torch.backends.mps.is_available():
-        devices.append(DeviceType.MPS)
-    return devices
-
-
-def _gpu_devices() -> list:
-    """Detect available GPU devices (MPS or CUDA)."""
-    return [d for d in _available_devices() if d != DeviceType.CPU]
-
-
-AVAILABLE_DEVICES = _available_devices()
-GPU_DEVICES = _gpu_devices()
-
-
-@pytest.fixture(params=AVAILABLE_DEVICES, ids=lambda d: f"device={d.value}")
-def any_device(request: pytest.FixtureRequest) -> DeviceType:
-    """Parameterize over all available devices (cpu, mps, cuda).
-
-    Use for Tier 1 and Tier 2 models that can run on any device.
-    """
-    return request.param
-
-
-@pytest.fixture(
-    params=GPU_DEVICES if GPU_DEVICES else [pytest.param("skip", marks=pytest.mark.skip("No GPU available"))],
-    ids=lambda d: f"device={d.value}" if isinstance(d, DeviceType) else "no-gpu",
-)
-def gpu_device(request: pytest.FixtureRequest) -> DeviceType:
-    """Parameterize over GPU-only devices (mps, cuda).
-
-    Use for Tier 3 models that are too slow on CPU.
-    Skips automatically if no GPU is available.
-    """
-    return request.param
